requirements.txt
playwright==1.40.0
pytest==7.4.0
pytest-xdist==3.3.0
pytest-html==3.2.0
pytest-logger==0.5.1
python-dotenv==1.0.0